+++ qcop/adapters/__init__.py
@@ -1,6 +1,7 @@
 """Must import all adapters here for them to be found by the AdapterRegistry."""
 
 from .base import *  # noqa: F403
+from .crest import CRESTAdapter  # noqa: F401
 from .file import FileAdapter  # noqa: F401
 from .geometric import GeometricAdapter  # noqa: F401
 from .qcengine import QCEngineAdapter  # noqa: F401